Modify formatQuantLabel so it attempts to always show something
As it stood, formatQuantLabel would simply show 0.0 for values in the
browser where the value was "too small". This meant that the upper and lower
values in the vertical "ruler" wouldn't be very helpful for tracks where the
min and max values were rather small.

This change attempts to retain the existing formatting for values that are
significant enough to show something sensible, but also work for smaller
values.

Also, because I'm not 100% sure that the value passed to formatQuantLabel
is *always* numeric, I've attempted to handle being passed a string
representation of a value too.